requirements.txt
asgiref
Django
djangorestframework
psycopg2-binary
pytz
sqlparse
django-filter
dj-rest-auth 
django-cors-headers
drf-yasg